refactor: add db core barrel exports
