+++ frontends/nextjs/src/lib/db/core/entities.ts
@@ -1,0 +1,22 @@
+// Domain re-exports
+export * from '../auth'
+export * from '../users'
+export * from '../credentials'
+export * from '../sessions'
+export * from '../workflows'
+export * from '../lua-scripts'
+export * from '../pages'
+export * from '../schemas'
+export * from '../comments'
+export * from '../app-config'
+export * from '../system-config'
+export * from '../components'
+export * from '../css-classes'
+export * from '../dropdown-configs'
+export * from '../tenants'
+export * from '../packages'
+export * from '../power-transfers'
+export * from '../smtp-config'
+export * from '../god-credentials'
+export * from '../database-admin'
+export * from '../error-logs'
